revert: drop stderr surfacing from this PR, keep the minimal snapshot CLI fix
The detached-launch error refactor pulled in secret-redaction concerns
that deserve their own focused change. The captured stderr already lands
in task_run_events details, which is sufficient for diagnosis. This PR
now contains only the option-mapping fix and its regression test.
